+++ CHANGELOG.md
@@ -9,13 +9,17 @@ nav_order: 1
 <!-- Always keep the following header in place: -->
 <!-- ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD -->
 
+## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
+
+- Add `disaster_recovery` service integration type support
+- Run client-side validation for `aiven_kafka_schema` AVRO type schema
+
 ## [4.27.0] - 2024-10-09
 
 - Remove `aiven_thanos` from beta resources
 - Removes `receiver_ingesting_remote_write_uri` and `store_uri` Thanos connection info fields
 - Adds `stringtype` to `flink_external_postgresql_user_config` service integration
 - Fix `terraform import` for services with additional disk space or read replica service integration
-- Run client-side validation for `aiven_kafka_schema` AVRO type schema
 
 ## [4.26.0] - 2024-09-25